feat: Turbopack typed links (#81528)
# Typed Links: Turbopack support

This PR replaces the old `experimental.typedRoutes` implementation with
a new approach that works on both Turbopack and Webpack.

_Note: a previous version of this PR contained new props for the `Link`
component: `path`, `params`, and `searchParams`. In the interest of
bringing typed links support to Turbopack as soon as possible, I've
removed that change from this PR and will create a new one later._