run config deprecation checks only on user-provided configuration (#82613)
The deprecation warning code is fairly brittle right now since there's a
mixture of calling things with `userConfig` and `result`. It was assumed
that `userConfig` was the actual user-provided configuration, but it
turns out it's actually the default configuration when a user
configuration isn't specified.

This moves all the deprecation checks to be before we actually do any
sort of merging. This restores the intent behind `userConfig` and will
prevent deprecation warnings from being logged due to defaults.

Alternative to #82593