Turbopack: fix edge function name (#82617)
Edge functions with route `/` had `"name": "/"` in the `middleware-manifest.json`. When deploying to Vercel, this wouldn't work as the Vercel builder turned that into `"name": ""`.
For the root function, Webpack emits a name like `"app/page"` or `"pages/index"`.
This aligns Turbopack with Webpack

I think the name itself can be pretty arbitrary. As long as the `_ENTRIES` access and the middleware manifests agree (which is covered by the regular e2e tests), and as long as it's not literally `"/"` which breaks on Vercel.

Closes PACK-5245