[turbopack] Remove a debug assertions in client references endpoint (#82725)
Remove a debug assertion code in the `GlobalBuildInformation` implementation that was validating client references for endpoints.  This was inserted as a sanity check that two traversals were doing the same thing, but this is overkill and can be trigged accidentally by an eventual consistency condition.  Just drop them.

If an application is edited during this computation it is possible that the module graph that GlobalBuildInformation has access to is different than the set of references that `find_server_entries` will identified, this transient condition could trigger the panic.