Route modules: Call setReferenceManifestsSingleton for app-route to match app-page (#82908)
## What?

The changes in #82903 highlighted some failing tests because `use cache`
relies on the global singleton being set which accidentally depended on
the `load-components` call to `setReferenceManifestsSingleton`. The
`app-page` template did the right thing calling
`setReferenceManifestsSingleton` already but `app-route` did not,
causing those tests to fail when I removed calling
`setReferenceManifestsSingleton` in the PR further in the stack.