[turbopack] Add support for partial glob matches (#82906)
### What

Add a `contains` option to our Glob API that is compatible with the [`picomatch`](https://www.npmjs.com/package/picomatch/v/4.0.3) `contains` option

### Why

This is the default behavior of the globs passed to NFT configuration options in next, and so we need this feature to be compatible in our rewrite.

### How

Just drop `$` and `^` from the generated regexes when `contains` is set.

For the nft globs we set `contains:true` which is compatible with the current behavior in [`collect-build-traces.ts`](https://github.com/vercel/next.js/blob/b56d1e1af1ad37f96e1905662b15ac5f4707726b/packages/next/src/build/collect-build-traces.ts#L189-L192)

Closes PACK-5305