Turbopack: Redefine Background backend jobs (#83174)
### What?

We don't need background jobs that run when idle. We already have a `idle_start` event for that.
Instead we need "background jobs" that do not count towards activeness of the system (do not contribute to the compile duration and these effects).

Also remove a bunch of unused stuff from the manager.