Turbopack: Skip manifest loading in server when using route modules (#82903)
## What?

When using Turbopack we'd try to load these files even when they don't
exist, which adds latency because of the retrying. This PR ensures the
files are skipped from loading where possible, which moves the place
where it starts loading the manifest to the route modules instead. The
exception to this is the static generation steps as those don't rely on
route modules yet. This is something still being worked on.