auto-enable errors.json merge driver (#83420)
The merge driver for `errors.json` has been around for a while, and
seems to work without issues. This PR enables it by default by
installing it in `postinstall`(via the existing `git-configure.mjs`
script which is invoked from there already)

It also adds `scripts/merge-errors-json/uninstall` to temporarily
disable it in case of issues. If the driver crashes, we'll also print
instructions on how to disable it.